i2 - add tests
