+++ .github/workflows/publish_crate.yml
@@ -1,0 +1,38 @@
+name: Publish Crate
+
+on:
+  push:
+    tags:
+      - "v*"
+
+jobs:
+  publish:
+    name: Publish to crates.io
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v3
+
+      - name: Setup Rust
+        uses: actions-rs/toolchain@v1
+        with:
+          profile: minimal
+          toolchain: stable
+          override: true
+
+      - name: Check version match
+        run: |
+          TAG_NAME=${GITHUB_REF#refs/tags/v}
+          CARGO_TOML_VERSION=$(grep '^version =' Cargo.toml | sed 's/version = "\(.*\)"/\1/')
+          if [ "$TAG_NAME" != "$CARGO_TOML_VERSION" ]; then
+            echo "Tag version ($TAG_NAME) does not match Cargo.toml version ($CARGO_TOML_VERSION)"
+            exit 1
+          fi
+
+      - name: Publish to crates.io
+        uses: actions-rs/cargo@v1
+        with:
+          command: publish
+          args: --token ${{ secrets.CRATES_IO_TOKEN }}
+        env:
+          CARGO_REGISTRY_TOKEN: ${{ secrets.CRATES_IO_TOKEN }}
