+++ tests/test_mdbook_reference_table/reference-table.yaml
@@ -1,6 +1,8 @@
 reference-table:
   - id: "audita-quot"
     path: "content/subject-1/audita-quot.md"
+  - id: "aliter-facta-cornibus-tandem"
+    path: "content/subject-1/aliter-facta-cornibus-tandem.md"
   - id: "bos-quae-inde-limen"
     path: "content/subject-2/subfolder-1/bos-quae-inde-limen.md"
   - id: "esse-corpora-achaidas-sacros-ad-quas"
