+++ tests/test_mdbook_reference_table/src/SUMMARY.md
@@ -3,5 +3,6 @@
 [Summary](./SUMMARY.md)
 
 - [Audita quot](./content/subject-1/audita-quot.md)
+- [Aliter facta cornibus tandem](./content/subject-1/aliter-facta-cornibus-tandem.md)
 - [Bos quae inde limen](./content/subject-2/subfolder-1/bos-quae-inde-limen.md)
 - [Esse corpora Achaidas sacros ad quas](./content/subject-3/subfolder-1/subsubfolder-1/esse-corpora-achaidas-sacros-ad-quas.md)
