+++ tests/test_mdbook_reference_table/src/content/subject-3/subfolder-1/subsubfolder-1/esse-corpora-achaidas-sacros-ad-quas.md
@@ -42,4 +42,5 @@ Constituis ut undis si ipsa rex *temeraria* arma gramina. **Iliaden frementis** 
 ## References
 
 - {{reference: {id: "audita-quot", title: "Audit quot"}}}
+- {{reference: {id: "aliter-facta-cornibus-tandem", title: "Aliter facta cornibus tandem"}}}
 - {{reference: {id: "bos-quae-inde-limen", title: "Bos quae inde lime"}}}
